# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "WMS Web API Service Order Documentation - Development",
    "version": "4.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "x-amazon-apigateway-binary-media-types": [
    "multipart/form-data"
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "api_key": {
        "type": "apiKey",
        "name": "x-api-key",
        "in": "header"
      }
    },
    "schemas": {
      "StockAvailabilityHistoryQueryDto": {
        "type": "object",
        "required": [
          "materialCode"
        ],
        "properties": {
          "customerCode": {
            "type": "string"
          },
          "warehouseCode": {
            "type": "string"
          },
          "materialCode": {
            "type": "string"
          }
        }
      },
      "StockAvailabilityListDto": {
        "type": "object",
        "properties": {
          "page": {
            "type": "integer",
            "minimum": 1,
            "example": 1
          },
          "limit": {
            "type": "integer",
            "minimum": 1,
            "maximum": 100,
            "example": 10
          },
          "search": {
            "type": "string",
            "example": "BRG-001"
          },
          "searchBy": {
            "type": "string",
            "enum": [
              "materialCode",
              "materialName",
              "materialBrand",
              "customerName",
              "warehouseName"
            ],
            "description": "Kolom search; tanpa searchBy = LIKE gabung 5 kolom (parity SP)"
          },
          "customerCode": {
            "type": "string",
            "description": "Filter exact"
          },
          "warehouseCode": {
            "type": "string",
            "description": "Filter exact"
          },
          "order": {
            "type": "string",
            "enum": [
              "customerCode",
              "customerName",
              "warehouseCode",
              "warehouseName",
              "materialCode",
              "materialName",
              "materialBrand",
              "qtySOH",
              "qtyPlanIncoming",
              "qtyPlanOutgoing",
              "createdAt"
            ]
          },
          "sort": {
            "type": "string",
            "enum": [
              "asc",
              "desc"
            ]
          }
        }
      }
    }
  },
  "paths": {
    "/v1/stock-availability": {
      "get": {
        "summary": "Q1 — List stock per material (paging, search LIKE, sort; parity usp_ListStockAvailability)",
        "tags": [
          "StockAvailability"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "api_key": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100
            }
          },
          {
            "in": "query",
            "name": "search",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "searchBy",
            "schema": {
              "type": "string",
              "enum": [
                "materialCode",
                "materialName",
                "materialBrand",
                "customerName",
                "warehouseName"
              ]
            }
          },
          {
            "in": "query",
            "name": "customerCode",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "warehouseCode",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "order",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "sort",
            "schema": {
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List stock availability"
          },
          "401": {
            "description": "Unauthorized"
          },
          "422": {
            "description": "Validation errors"
          }
        },
        "x-amazon-apigateway-integration": {
          "httpMethod": "get",
          "uri": "http://localhost:3000/v1/stock-availability",
          "connectionType": "VPC_LINK",
          "connectionId": "tbr2om",
          "passthroughBehavior": "when_no_match",
          "type": "http_proxy"
        }
      },
      "options": {
        "responses": {
          "200": {
            "description": "200 response",
            "headers": {
              "Access-Control-Allow-Origin": {
                "schema": {
                  "type": "string"
                }
              },
              "Access-Control-Allow-Methods": {
                "schema": {
                  "type": "string"
                }
              },
              "Access-Control-Allow-Headers": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "x-amazon-apigateway-integration": {
          "responses": {
            "default": {
              "statusCode": "200",
              "responseParameters": {
                "method.response.header.Access-Control-Allow-Methods": "'DELETE,GET,OPTIONS,PUT,POST,PATCH'",
                "method.response.header.Access-Control-Allow-Headers": "'Content-Type,X-Amz-Date,Authorization,X-Api-Key,X-Amz-Security-Token,User-Agent,Lang'",
                "method.response.header.Access-Control-Allow-Origin": "'*'"
              }
            }
          },
          "requestTemplates": {
            "application/json": "{\"statusCode\": 200}"
          },
          "passthroughBehavior": "when_no_match",
          "type": "mock"
        }
      }
    },
    "/v1/stock-availability/history": {
      "get": {
        "summary": "Q2 — Riwayat mutasi stock on hand satu material (parity usp_GetHistoryStockOnHand)",
        "tags": [
          "StockAvailability"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "api_key": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "customerCode",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "warehouseCode",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "materialCode",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "History rows"
          },
          "422": {
            "description": "Validation errors"
          }
        },
        "x-amazon-apigateway-integration": {
          "httpMethod": "get",
          "uri": "http://localhost:3000/v1/stock-availability/history",
          "connectionType": "VPC_LINK",
          "connectionId": "tbr2om",
          "passthroughBehavior": "when_no_match",
          "type": "http_proxy"
        }
      },
      "options": {
        "responses": {
          "200": {
            "description": "200 response",
            "headers": {
              "Access-Control-Allow-Origin": {
                "schema": {
                  "type": "string"
                }
              },
              "Access-Control-Allow-Methods": {
                "schema": {
                  "type": "string"
                }
              },
              "Access-Control-Allow-Headers": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "x-amazon-apigateway-integration": {
          "responses": {
            "default": {
              "statusCode": "200",
              "responseParameters": {
                "method.response.header.Access-Control-Allow-Methods": "'DELETE,GET,OPTIONS,PUT,POST,PATCH'",
                "method.response.header.Access-Control-Allow-Headers": "'Content-Type,X-Amz-Date,Authorization,X-Api-Key,X-Amz-Security-Token,User-Agent,Lang'",
                "method.response.header.Access-Control-Allow-Origin": "'*'"
              }
            }
          },
          "requestTemplates": {
            "application/json": "{\"statusCode\": 200}"
          },
          "passthroughBehavior": "when_no_match",
          "type": "mock"
        }
      }
    }
  },
  "tags": [
    {
      "name": "StockAvailability",
      "description": "Dashboard stock availability (parity CoreApp StockAvailabilityDashboard)"
    }
  ]
}